requirements.txt
wheel~=0.41.2
Django~=5.1.3
asgiref~=3.8.1
sqlparse~=0.5.2
djangorestframework~=3.15.2
factory-boy~=3.3.1
setuptools~=70.0.0
environ~=1.0